remove unused scan script
